refactor: move filename parsing to ArchiverModels
- move Document.parseFilename, DateParser and DateFormatter.yyyyMMdd from Shared
- move filename placeholders to Document statics
- update call sites in ArchiverStore, ArchiverDocumentProcessing and tests
