+++ ArchiverLib/Sources/ArchiverModels/DateParser.swift
@@ -6,9 +6,12 @@
 //
 
 import Foundation
+import OSLog
 
 /// Parse several kinds of dates in a String.
-nonisolated public enum DateParser: Log {
+nonisolated public enum DateParser {
+
+    private static let logger = Logger(subsystem: Bundle.main.bundleIdentifier ?? "de.JulianKahnert.PDFArchiveViewer", category: "date-parser")
 
     private struct ParserResult: Codable {
         let date: Date
@@ -46,7 +49,8 @@ nonisolated public enum DateParser: Log {
     private static func localParse(_ raw: String) async -> [ParserResult] {
         let types: NSTextCheckingResult.CheckingType = .date
         guard let detector = try? NSDataDetector(types: types.rawValue) else {
-            Self.log.criticalAndAssert("Could not create NSDataDetector")
+            logger.critical("Could not create NSDataDetector")
+            assertionFailure("Could not create NSDataDetector")
             return []
         }
