+++ ArchiverLib/Tests/ArchiverStoreTests/ArchiveBackendTests/DocumentTests.swift
@@ -285,7 +285,7 @@ struct DocumentTests {
         // setup
         let date = try #require(dateFormatter.date(from: "2018-05-12"))
         let document = Document(id: 1,
-                                url: URL(fileURLWithPath: "~/Downloads/2018-05-12--\(Constants.documentDescriptionPlaceholder)__\(Constants.documentTagPlaceholder).pdf"),
+                                url: URL(fileURLWithPath: "~/Downloads/2018-05-12--\(Document.descriptionPlaceholder)__\(Document.tagPlaceholder).pdf"),
                                 date: date,
                                 specification: "",
                                 tags: [],
